hip_prof_(api/str).h moving under hcc_detail dir
